Update AggregationServer.java
Aggregation server has been designed to stay current and will remove any items in the JSON that have come from content servers which it has not communicated with for 30 seconds. 

The server takes the file in hashmap along with data and timestamp to create weather data and the data is transformed again to json format when get request is called.